ci(test_espefuse): Migrate tests from unittest to pytest
