+++ test/efuse_scripts/esp32/execute_efuse_script.py
