+++ test/efuse_scripts/esp32xx/execute_efuse_script2.py
